Use explicit commands rather than environment variables in Cirrus
For some reason, the environment variables were being set in the top
and could be used in top-level nodes, but they were failing in matrix
tests:
https://github.com/jakewilliami/local-ip-address/runs/9611248995

This commit expands all of these out into explicit commmands, hence
doing away with most variables.  I have kept TOOLCHAIN defined as the
MSRV in the hope that this works.