Use individual tasks over one task with matrix in Cirrus CI
One question I had when researching Cirrus CI for our use case is
whether it is better to have individual tasks which run each separate
test (but thus require to download and build Rust for each test), or
to have one task which builds Rust only once, and can then run
subtasks using the matrix keyword.  While I had not yet found a
consensus in my research, it seems The Gods have decided for us, as
the environment being set up at the top level of the Cirrus CI config
did not allow the subtasks to have the same set up.  This commit
hopefully fixes this issue, but using individual tasks over one task
with a matrix.  It will, however, download and build Rust for each test.