Correct off-by-one error in parsing of getifaddrs for BSD
We previously used a while loop with a condition to iterate over the
results of libc::gerifaddrs, however as we checked if ifa.ifa_next was
null from the start, we were not checking one network interface.  This
caused somme bugs in tests for FreeBSD, which we partially fix in the
present commit (there is another bug that will be fixed in the next
commit, involving a hard-coded search for specific interface names).