Search for epair0b as well as en0 interfaces from FreeBSD
FreeBSD tests were failing due to no en0 interfaces.  In this commit,
we modify the find_ifa function: this method now is called find_ifas,
and has a new type signature, in which it accepts a vector of
possible interface names on which to match.

For future, it would be ideal to dynamically determine whether the
interface without hard-coding interface names.